requirements.txt
requests>=2.25.0
pyyaml>=5.1
sqlite-utils>=3.0
python-dateutil>=2.8.1
click>=7.1.2
pandas>=1.0.0
plotly>=4.14.0
numpy>=1.19.0